+++ src/main/java/workwithzoo/sourceofprofit/Buyable.java
@@ -2,13 +2,21 @@
 package workwithzoo.sourceofprofit;
 
 /**
- *
+ * For instance, who can to buys and sells.
  * @author DantalioNxxi
  */
 public interface Buyable {
     
-    double sell();
+    /**
+     * Sell object.
+     * @return cost of the selling object
+     */
+    float sell();
     
+    /**
+     * Buy object.
+     * @return instance of the Object.
+     */
     Buyable buy();
     
 }
